requirement_GSFF.txt
Package                   Version
------------------------- --------------
absl-py                   2.0.0
aiohttp                   3.8.5
aiosignal                 1.3.1
alembic                   1.14.0
annotated-types           0.7.0
antlr4-python3-runtime    4.9.3
anyio                     4.9.0
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 3.0.0
async-lru                 2.0.5
async-timeout             4.0.3
attrs                     23.1.0
babel                     2.17.0
beautifulsoup4            4.13.3
bleach                    6.2.0
Brotli                    1.0.9
cachetools                5.3.1
certifi                   2022.12.7
cffi                      1.17.1
cftime                    1.6.4
charset-normalizer        2.1.1
click                     8.1.8
cmake                     3.27.6
colorama                  0.4.6
colorlog                  6.9.0
comm                      0.2.2
configs                   3.0.3
contourpy                 1.1.1
cycler                    0.12.0
debugpy                   1.8.13
decorator                 5.1.1
defusedxml                0.7.1
dhg                       0.9.4
docker-pycreds            0.4.0
einops                    0.7.0
et_xmlfile                2.0.0
eval_type_backport        0.2.2
exceptiongroup            1.2.2
executing                 2.2.0
fastjsonschema            2.21.1
filelock                  3.13.1
fonttools                 4.43.0
fqdn                      1.5.1
frozenlist                1.4.0
fsspec                    2023.9.2
gitdb                     4.0.12
GitPython                 3.1.44
gmpy2                     2.1.2
google-auth               2.23.2
google-auth-oauthlib      1.0.0
greenlet                  3.1.1
grpcio                    1.59.0
h11                       0.14.0
h5py                      3.9.0
httpcore                  1.0.7
httpx                     0.28.1
huggingface_hub           0.24.6
hydra-core                1.3.2
idna                      3.4
importlib_metadata        7.2.0
importlib_resources       6.4.0
ipykernel                 6.29.5
ipython                   8.18.1
isoduration               20.11.0
jedi                      0.19.2
Jinja2                    3.1.3
joblib                    1.3.2
json5                     0.10.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
jupyter_client            8.6.3
jupyter_core              5.7.2
jupyter-events            0.12.0
jupyter-lsp               2.2.5
jupyter_server            2.15.0
jupyter_server_terminals  0.5.3
jupyterlab                4.3.6
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
kiwisolver                1.4.5
lazy_loader               0.4
lightning-utilities       0.9.0
lit                       17.0.1
llvmlite                  0.43.0
Mako                      1.3.8
Markdown                  3.4.4
MarkupSafe                2.1.5
matplotlib                3.8.0
matplotlib-inline         0.1.7
mistune                   3.1.2
mkl_fft                   1.3.11
mkl_random                1.2.8
mkl-service               2.4.0
mne                       1.7.1
mne-connectivity          0.7.0
mne-features              0.3
mpmath                    1.3.0
multidict                 6.0.4
nbclient                  0.10.2
nbconvert                 7.16.6
nbformat                  5.10.4
nest-asyncio              1.6.0
netCDF4                   1.7.1.post1
networkx                  3.2.1
notebook                  7.3.3
notebook_shim             0.2.4
numba                     0.60.0
numpy                     1.24.4
oauthlib                  3.2.2
omegaconf                 2.3.0
opencv-python             4.10.0.84
openpyxl                  3.1.5
optuna                    4.1.0
overrides                 7.7.0
packaging                 23.2
pandas                    2.2.2
pandocfilters             1.5.1
parso                     0.8.4
pillow                    10.2.0
pip                       24.0
platformdirs              3.10.0
pooch                     1.7.0
prometheus_client         0.21.1
prompt_toolkit            3.0.50
protobuf                  4.24.3
psutil                    6.1.1
pure_eval                 0.2.3
pyasn1                    0.5.0
pyasn1-modules            0.3.0
pycparser                 2.22
pydantic                  2.10.4
pydantic_core             2.27.2
pyEDFlib                  0.1.34
Pygments                  2.19.1
pyparsing                 3.1.1
PySocks                   1.7.1
python-dateutil           2.8.2
python-json-logger        3.3.0
pytorch-lightning         2.0.9.post0
pytz                      2024.1
PyWavelets                1.6.0
pywin32                   310
pywinpty                  2.0.15
pyxdf                     1.16.4
PyYAML                    6.0.1
pyzmq                     26.3.0
referencing               0.36.2
requests                  2.32.3
requests-oauthlib         1.3.1
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.23.1
rsa                       4.9
safetensors               0.4.5
scikit-learn              1.3.1
scipy                     1.11.3
seaborn                   0.13.2
Send2Trash                1.8.3
sentry-sdk                2.19.2
setproctitle              1.3.4
setuptools                69.5.1
six                       1.16.0
smmap                     5.0.2
sniffio                   1.3.1
soupsieve                 2.6
SQLAlchemy                2.0.37
stack-data                0.6.3
sympy                     1.13.1
tensorboard               2.14.1
tensorboard-data-server   0.7.1
tensorboardX              2.6.2.2
terminado                 0.18.1
threadpoolctl             3.2.0
timm                      1.0.12
tinycss2                  1.4.0
tomli                     2.2.1
torch                     1.13.1+cu116
torch-geometric           2.6.1
torchaudio                0.13.1+cu116
torchmetrics              1.2.0
torchnet                  0.0.4
torchsummary              1.5.1
torchvision               0.14.1+cu116
tornado                   6.4.1
tqdm                      4.66.1
traitlets                 5.14.3
types-python-dateutil     2.9.0.20241206
typing_extensions         4.9.0
tzdata                    2024.1
uri-template              1.3.0
urllib3                   1.26.13
visdom                    0.2.4
wandb                     0.19.1
wcwidth                   0.2.13
webcolors                 24.11.1
webencodings              0.5.1
websocket-client          1.8.0
Werkzeug                  3.0.0
wheel                     0.43.0
win-inet-pton             1.1.0
xarray                    2024.6.0
yarl                      1.9.2
zipp                      3.19.2
